separate the process functions in their own module
